Move test cases to their own directory to avoid conflict
